+++ src/PSRule.Rules.Azure/rules/Azure.Redis.Rule.ps1
@@ -124,7 +124,7 @@ Rule 'Azure.Redis.Version' -Ref 'AZR-000347' -Type 'Microsoft.Cache/redis' -Tag 
 }
 
 # Synopsis: Azure Cache for Redis is being retired. Migrate to Azure Managed Redis.
-Rule 'Azure.Redis.Retirement' -Ref 'AZR-000506' -Type 'Microsoft.Cache/redis' -Tag @{ release = 'GA'; ruleSet = '2025_12'; 'Azure.WAF/pillar' = 'Operational Excellence'; } {
+Rule 'Azure.Redis.MigrateAMR' -Ref 'AZR-000506' -Type 'Microsoft.Cache/redis' -Tag @{ release = 'GA'; ruleSet = '2025_12'; 'Azure.WAF/pillar' = 'Operational Excellence'; } {
     $Assert.Fail($LocalizedData.AzureCacheRedisRetirement)
 }